refactoring all ids to long and handled soft delete of products/categories and there side effects on Highlight/CartDetail/OrderDetail

+++ src/main/java/com/ecommerce/basic/models/CartItem.java
@@ -21,7 +21,7 @@ import java.util.List;
 public class CartItem {
 	@Id
 	@GeneratedValue
-	private int cartId;
+	private Long cartId;
 
 	@JsonIgnore
 	@ToString.Exclude
